+++ engine/context/src/main/java/io/deephaven/engine/context/EmptyQueryScope.java
@@ -3,8 +3,13 @@
  */
 package io.deephaven.engine.context;
 
+import io.deephaven.engine.liveness.LivenessReferent;
+import org.jetbrains.annotations.NotNull;
+
+import java.lang.ref.WeakReference;
 import java.util.Collections;
 import java.util.Set;
+import java.util.stream.Stream;
 
 public class EmptyQueryScope implements QueryScope {
     public final static EmptyQueryScope INSTANCE = new EmptyQueryScope();
@@ -39,5 +44,35 @@ public class EmptyQueryScope implements QueryScope {
     @Override
     public <T> void putParam(String name, T value) {
         throw new IllegalStateException("EmptyQueryScope cannot create parameters");
+    }
+
+    @Override
+    public boolean tryManage(@NotNull LivenessReferent referent) {
+        throw new UnsupportedOperationException("tryManage");
+    }
+
+    @Override
+    public boolean tryUnmanage(@NotNull LivenessReferent referent) {
+        throw new UnsupportedOperationException("tryUnmanage");
+    }
+
+    @Override
+    public boolean tryUnmanage(@NotNull Stream<? extends LivenessReferent> referents) {
+        throw new UnsupportedOperationException("tryUnmanage");
+    }
+
+    @Override
+    public boolean tryRetainReference() {
+        throw new UnsupportedOperationException("tryRetainReference");
+    }
+
+    @Override
+    public void dropReference() {
+        throw new UnsupportedOperationException("dropReference");
+    }
+
+    @Override
+    public WeakReference<? extends LivenessReferent> getWeakReference() {
+        throw new UnsupportedOperationException("getWeakReference");
     }
 }
