+++ engine/context/src/main/java/io/deephaven/engine/context/PoisonedQueryScope.java
@@ -3,9 +3,13 @@
  */
 package io.deephaven.engine.context;
 
+import io.deephaven.engine.liveness.LivenessReferent;
 import io.deephaven.util.ExecutionContextRegistrationException;
+import org.jetbrains.annotations.NotNull;
 
+import java.lang.ref.WeakReference;
 import java.util.Set;
+import java.util.stream.Stream;
 
 public class PoisonedQueryScope implements QueryScope {
 
@@ -45,5 +49,35 @@ public class PoisonedQueryScope implements QueryScope {
     @Override
     public <T> void putParam(String name, T value) {
         fail();
+    }
+
+    @Override
+    public boolean tryManage(@NotNull LivenessReferent referent) {
+        return fail();
+    }
+
+    @Override
+    public boolean tryUnmanage(@NotNull LivenessReferent referent) {
+        return fail();
+    }
+
+    @Override
+    public boolean tryUnmanage(@NotNull Stream<? extends LivenessReferent> referents) {
+        return fail();
+    }
+
+    @Override
+    public boolean tryRetainReference() {
+        return fail();
+    }
+
+    @Override
+    public void dropReference() {
+        fail();
+    }
+
+    @Override
+    public WeakReference<? extends LivenessReferent> getWeakReference() {
+        return fail();
     }
 }
