+++ engine/context/src/main/java/io/deephaven/engine/context/QueryScope.java
@@ -3,24 +3,19 @@
  */
 package io.deephaven.engine.context;
 
-import io.deephaven.time.DateTimeUtils;
-import io.deephaven.hash.KeyedObjectHashMap;
-import io.deephaven.hash.KeyedObjectKey;
+import io.deephaven.engine.liveness.LivenessNode;
 import io.deephaven.base.log.LogOutput;
 import io.deephaven.base.log.LogOutputAppendable;
-import io.deephaven.api.util.NameValidator;
-import io.deephaven.util.QueryConstants;
 import org.jetbrains.annotations.NotNull;
 
-import java.time.Duration;
-import java.time.Instant;
-import java.time.Period;
-import java.util.*;
+import java.util.Collection;
+import java.util.Objects;
+import java.util.Set;
 
 /**
  * Variable scope used to resolve parameter values during query execution.
  */
-public interface QueryScope extends LogOutputAppendable {
+public interface QueryScope extends LivenessNode, LogOutputAppendable {
 
     /**
      * Adds a parameter to the default instance {@link QueryScope}, or updates the value of an existing parameter.
@@ -66,47 +61,6 @@ public interface QueryScope extends LogOutputAppendable {
         private MissingVariableException(final Throwable cause) {
             super(cause);
         }
-    }
-
-    /**
-     * Apply conversions to certain scope variable values.
-     *
-     * @param value value
-     * @return value, or an appropriately converted substitute.
-     */
-    private static Object applyValueConversions(final Object value) {
-        if (value instanceof String) {
-            final String stringValue = (String) value;
-
-            if (stringValue.length() > 0 && stringValue.charAt(0) == '\''
-                    && stringValue.charAt(stringValue.length() - 1) == '\'') {
-                final String datetimeString = stringValue.substring(1, stringValue.length() - 1);
-
-                final Instant instant = DateTimeUtils.parseInstantQuiet(datetimeString);
-                if (instant != null) {
-                    return instant;
-                }
-
-                final long localTime = DateTimeUtils.parseDurationNanosQuiet(datetimeString);
-                if (localTime != QueryConstants.NULL_LONG) {
-                    return localTime;
-                }
-
-                final Period period = DateTimeUtils.parsePeriodQuiet(datetimeString);
-                if (period != null) {
-                    return period;
-                }
-
-                final Duration duration = DateTimeUtils.parseDurationQuiet(datetimeString);
-                if (duration != null) {
-                    return duration;
-                }
-
-                throw new RuntimeException("Cannot parse datetime/time/period : " + stringValue);
-            }
-        }
-
-        return value;
     }
 
     // -----------------------------------------------------------------------------------------------------------------
@@ -176,7 +130,7 @@ public interface QueryScope extends LogOutputAppendable {
     <T> T readParamValue(final String name, final T defaultValue);
 
     /**
-     * Add a parameter to the scope.
+     * Add a parameter to the scope. Objects passed in will have their liveness managed by the scope.
      *
      * @param name parameter name.
      * @param value parameter value.
@@ -214,115 +168,5 @@ public interface QueryScope extends LogOutputAppendable {
             }
         }
         return logOutput.nl().append('}');
-    }
-
-    // -----------------------------------------------------------------------------------------------------------------
-    // Map-based implementation, with remote scope and object reflection support
-    // -----------------------------------------------------------------------------------------------------------------
-
-    class StandaloneQueryScope implements QueryScope {
-
-        private final KeyedObjectHashMap<String, ValueRetriever> valueRetrievers =
-                new KeyedObjectHashMap<>(new ValueRetrieverNameKey());
-
-        @Override
-        public Set<String> getParamNames() {
-            return valueRetrievers.keySet();
-        }
-
-        @Override
-        public boolean hasParamName(String name) {
-            return valueRetrievers.containsKey(name);
-        }
-
-        @Override
-        public <T> QueryScopeParam<T> createParam(final String name) throws MissingVariableException {
-            // noinspection unchecked
-            final ValueRetriever<T> valueRetriever = valueRetrievers.get(name);
-            if (valueRetriever == null) {
-                throw new MissingVariableException("Missing variable " + name);
-            }
-            return valueRetriever.createParam();
-        }
-
-        @Override
-        public <T> T readParamValue(final String name) throws MissingVariableException {
-            // noinspection unchecked
-            final ValueRetriever<T> valueRetriever = valueRetrievers.get(name);
-            if (valueRetriever == null) {
-                throw new MissingVariableException("Missing variable " + name);
-            }
-            return valueRetriever.getValue();
-        }
-
-        @Override
-        public <T> T readParamValue(final String name, final T defaultValue) {
-            // noinspection unchecked
-            final ValueRetriever<T> valueRetriever = valueRetrievers.get(name);
-            if (valueRetriever == null) {
-                return defaultValue;
-            }
-            return valueRetriever.getValue();
-        }
-
-        @Override
-        public <T> void putParam(final String name, final T value) {
-            NameValidator.validateQueryParameterName(name);
-            // TODO: Can I get rid of this applyValueConversions? It's too inconsistent to feel safe.
-            valueRetrievers.put(name, new SimpleValueRetriever<>(name, applyValueConversions(value)));
-        }
-
-        private static abstract class ValueRetriever<T> {
-
-            private final String name;
-
-            protected ValueRetriever(String name) {
-                this.name = name;
-            }
-
-            public String getName() {
-                return name;
-            }
-
-            public abstract T getValue();
-
-            public abstract Class<T> getType();
-
-            public abstract QueryScopeParam<T> createParam();
-        }
-
-        private static class ValueRetrieverNameKey extends KeyedObjectKey.Basic<String, ValueRetriever> {
-
-            @Override
-            public String getKey(ValueRetriever valueRetriever) {
-                return valueRetriever.getName();
-            }
-        }
-
-        private static class SimpleValueRetriever<T> extends ValueRetriever<T> {
-
-            private final T value;
-
-            public SimpleValueRetriever(final String name, final T value) {
-                super(name);
-                this.value = value;
-            }
-
-            @Override
-            public T getValue() {
-                return value;
-            }
-
-            @Override
-            public Class<T> getType() {
-                // noinspection unchecked
-                return (Class<T>) (value != null ? value.getClass() : Object.class);
-            }
-
-            @Override
-            public QueryScopeParam<T> createParam() {
-                return new QueryScopeParam<>(getName(), getValue());
-            }
-        }
     }
 }
